pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ---------------------------------------------------------------------------
# Package metadata
# ---------------------------------------------------------------------------
[project]
name = "ai-context-cli"
version = "0.1.0"
description = "CLI: fetch an HTTP(S) URL, extract readable article HTML (Readability), emit Markdown; optional LLM summary via LiteLLM. Local file sources are planned (v1.x)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "CAO Anh (Lam)" }]
keywords = [
    "cli",
    "ai",
    "markdown",
    "scraper",
    "llm",
    "context",
    "readability",
    "litellm",
    "ollama",
    "openrouter",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Text Processing :: Markup",
    "Typing :: Typed",
]

dependencies = [
    "typer[all]>=0.12",        # [all] pulls in rich + shellingham
    "rich>=13.7",
    "pydantic>=2.7",
    "httpx>=0.27",
    "readability-lxml>=0.8",
    "markdownify>=0.12",
    "pyyaml>=6.0",
    "litellm>=1.40",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/Hushout/ai-context-cli"
Repository = "https://github.com/Hushout/ai-context-cli"
Issues = "https://github.com/Hushout/ai-context-cli/issues"
Documentation = "https://github.com/Hushout/ai-context-cli#readme"

[project.scripts]
ai-context-cli = "ai_context_cli.interfaces.cli:run_app"

# ---------------------------------------------------------------------------
# Optional / dev dependencies
# ---------------------------------------------------------------------------
[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-httpx>=0.30",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

# ---------------------------------------------------------------------------
# Build — src layout
# ---------------------------------------------------------------------------
[tool.hatch.build.targets.wheel]
packages = ["src/ai_context_cli"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers -q"

[tool.pytest.ini_options.markers]
e2e = "end-to-end tests that require network access (deactivated in CI)"

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["src/ai_context_cli"]
omit = ["src/ai_context_cli/interfaces/*"]  # CLI wiring excluded from coverage target

[tool.coverage.report]
fail_under = 80
show_missing = true

# ---------------------------------------------------------------------------
# Ruff (linter + formatter)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN"]
ignore = [
    "ANN101",   # missing type annotation for self
    "ANN102",   # missing type annotation for cls
]

[tool.ruff.lint.isort]
known-first-party = ["ai_context_cli"]

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["readability", "markdownify", "litellm"]
ignore_missing_imports = true
